Finished parsing

+++ Evaluation/Evaluation/Result.cs
@@ -2,11 +2,11 @@
 {
     public class Result
     {
-        public decimal ObjValue { get; }
+        public double ObjValue { get; }
         public ulong Time { get; }
         public bool IsFeasible { get; }
 
-        public Result(decimal objValue, ulong time, bool isFeasible)
+        public Result(double objValue, ulong time, bool isFeasible)
         {
             ObjValue = objValue;
             Time = time;
